ISSUE #4625 - Moved the queue logic into a separate file
