+++ frontend/src/v5/store/containers/containers.sagas.ts
@@ -40,66 +40,9 @@ import { selectContainerById, selectContainers, selectIsListPending } from './co
 import { compByColum } from '../store.helpers';
 import { getSortingFunction } from '@components/dashboard/dashboardList/useOrderedList/useOrderedList.helpers';
 import { SortingDirection } from '@components/dashboard/dashboardList/dashboardList.types';
+import { LifoQueue } from '@/v5/helpers/functions.helpers';
 
-type QueueItem<T> = { promise: Promise<T>, resolve, args };
-class Queue<T> {
-	private queue: QueueItem<T>[] = [] ;
-
-	private dict = {};
-
-	private batchSize;
-
-	private running = false;
-
-	private func: (...args) => Promise<T>;
-
-	private getPromise(args):QueueItem<T>  {
-		let prom = this.dict[JSON.stringify(args)];
-		if (!prom) {
-			prom = {};
-			prom.promise = new Promise((resolve) => prom.resolve = resolve);
-			prom.args = args;
-		}
-		
-		return prom;
-	}
-
-	private async runqueue() {
-		this.running = true;
-		while (this.queue.length) {
-			const batch =  this.queue.splice(Math.max(this.queue.length - this.batchSize, 0));
-			await Promise.all(batch.reverse().map(async (p) => {
-				console.log('fetching ' + p.args[p.args.length - 1]);
-				p.resolve(await this.func(...p.args));
-			}));
-		}
-		this.clearQueue();
-		this.running = false;
-	}
-
-	public enqueue(...args): Promise<T> {
-		const prom = this.getPromise(args);
-		this.queue.push(prom);
-
-		if (this.queue.length && !this.running) {
-			this.runqueue();
-		}
-
-		return prom.promise;
-	}
-
-	public clearQueue() {
-		this.queue = [];
-		this.dict = {};
-	}
-
-	public constructor(func: (...args) => Promise<T>, batchSize) {
-		this.func = func;
-		this.batchSize = batchSize;
-	}
-}
-
-const statsQueue = new Queue<ContainerStats>(API.Containers.fetchContainerStats, 30);
+const statsQueue = new LifoQueue<ContainerStats>(API.Containers.fetchContainerStats, 30);
 
 export function* addFavourites({ containerId, teamspace, projectId }: AddFavouriteAction) {
 	try {
